Fix problems with size of integer arguments to runtime functions.
When using large array descriptors, reduce_descriptor expects the dim argument
to be an 8-byte integer but the compiler generates code to pass in a 4-byte
integer. When that value is interpreted as 8 bytes long, the top 4 are non-zero
so it looks like a huge number.

The same problem occurs with the dim parameter to spread_descriptor and the
shift and dim parameters to the eoshift and cshift routines.

The conversion is done with convert_int(), which was previously named
_convert_int() and static in ast.c.