Fix implementation of bind(C, name='')
Use the same binding name as we would without the bind(C), so include the
module name in it. This is done by delaying processing of an explicit empty
bind name until routine process_bind to allow that case to be treated
differently from the implicit empty name case.

Also rearrange the code in routine semant1 to reduce the arbitrary
differences in the cases that process the two main bind attribute
reductions.  The processing for the unnamed variant is a subset of
the processing for the named variant, which is now more apparent
in the code.