Fix spurious internal error on call to IDATE.
Upstream code in routine ref_pd_subr in file semfunc.c changes
a call to the IDATE intrinsic into a user routine call if the
number or type of call arguments are invalid for the intrinsic.
The assert in routine rewrite_func_ast that is triggered due
to a missing definition of nargs is therefore unnecessary,
since the code is only reached when the call arguments have
already been validated.  Deleting the code with the asserts
leaves code for IDATE that is similar to nearby cases for other
intrinsics.