Fix bug passing contiguous pointer to array section.
This bug occurs when we have a contiguous pointer to an array section that
is passed to an assumed-shape formal like in this example:
	integer, pointer, contiguous :: p(:,:)
	integer, pointer :: matrix(:,:,:)
	allocate(matrix(...))
	p => matrix(:,:,n)
	call sub(p)

The section descriptor that we create and pass in to the procedure has an
lbase value that represents the distance from the start of the array to the
start of the section that the pointer represents.

When we pass in the address of the first element that p points to, that
distance is included in the pointer. So combined with the lbase offset in
the descriptor we end up indexing too far into the array (or past the end).

To fix this we need to either adjust the descriptor or pass in the address
of the array instead of the first element of the section. This fix does the
latter.