Enable nodepchk for simd clause or simd construct
Apply nodepchk to a loop/routine as a pragma nodepchk for simd
clause or simd construct. Since the host function has
information about nodepcheck, need to propagate the info to
outlined function when it was created because we lose sptr of
host function when we expand outlined function.

# Merged CL 142611
apply nodepchk to a loop/routine as a pragma nodepchk for simd clause or simd construct.

# Merged CL 142611
apply nodepchk to a loop/routine as a pragma nodepchk for simd clause or simd construct.

# Merged CL 142654
Additonal change nodepchk for C:
  Since the host function has information about nodepcheck.  Need to propagate the info to outlined function when it was created because we lose sptr of host function when we expand outlined function.

# Merged CL 142611
apply nodepchk to a loop/routine as a pragma nodepchk for simd clause or simd construct.

# Merged CL 142611
apply nodepchk to a loop/routine as a pragma nodepchk for simd clause or simd construct.

# Merged CL 142654
Additonal change nodepchk for C:
  Since the host function has information about nodepcheck.  Need to propagate the info to outlined function when it was created because we lose sptr of host function when we expand outlined function.

# Merged CL 142611
apply nodepchk to a loop/routine as a pragma nodepchk for simd clause or simd construct.

# Merged CL 142654
Additonal change nodepchk for C:
  Since the host function has information about nodepcheck.  Need to propagate the info to outlined function when it was created because we lose sptr of host function when we expand outlined function.

# Merged CL 143580
Add function prototype for apply_nodepchk to quiet warning
   in Flang.