Fix a bug in handling adjustl, adjustr, trim intrinsics
Consider invocations like
   foo()
     Character (max_charlen), Intent (Out) :: token(max_data_vals)
     token(1:n) = adjustl(token(1:n)

   bar()
     Character (*), Intent (In)    :: buffin
     buffall = adjustl(buffin)

The former gets converted to a subroutine call form, whereas
the latter remains as a func, the hash table would have an
entry for adjustl as void type, whereas the second call would
be expecting a return type of INT (actual result is passed as
hidden argument). This conflict type leads to an error. The
conversion to a subroutine call was enabled to prevent memory
allocation for a temp array used to store the return value of
the intrinsic, but the temp would never be used. This change
uses a single temp (in place of an array) to store the return
value of the call, thereby eliminating the need to convert it
to a subroutine call.

This change also fixes similar handling of 'adjustr' and 'trim'.

# Merged CL 143122
trim() intrinsic function was being set to A_FUNC type instead of A_INTR in CL#88505.
This fix, corrects that. The fix is not a result of any failure, rather a bug which was discovered
through FS#23495 and FS#19712.

THE BUG: When trim is used inside a forall() it allocates memory for a temp array to store return value
which is not necessary (this is not the actual return value, rather an integer return by the implementation
of trim in our runtime, the actual return value is a hidden argument).

   On setting the type to A_INTR, it follows the same path that adjust(l/r) go through in the 'convert forall' pass.
   The return value inside the	loop gets stored into a single integer temp which does not need memory to be allocated,
   thereby saving time and memory.