Fix a bug in handling adjustl, adjustr, trim intrinsics
Consider invocations like
   foo()
     Character (max_charlen), Intent (Out) :: token(max_data_vals)
     token(1:n) = adjustl(token(1:n)

   bar()
     Character (*), Intent (In)    :: buffin
     buffall = adjustl(buffin)

The former gets converted to a subroutine call form, whereas
the latter remains as a func, the hash table would have an
entry for adjustl as void type, whereas the second call would
be expecting a return type of INT (actual result is passed as
hidden argument). This conflict type leads to an error. The
conversion to a subroutine call was enabled to prevent memory
allocation for a temp array used to store the return value of
the intrinsic, but the temp would never be used. This change
uses a single temp (in place of an array) to store the return
value of the call, thereby eliminating the need to convert it
to a subroutine call.

This change also fixes similar handling of 'adjustr' and 'trim'.