Fix handling of TRIM intrinsic function
trim() intrinsic function was being set to A_FUNC type
instead of A_INTR.

Before, when trim is used inside a forall() it allocated
memory for a temp array to store return value which is not
necessary (this is not the actual return value, rather an
integer return by the implementation of trim in our runtime,
the actual return value is a hidden argument).

On setting the type to A_INTR, it follows the same path
that adjust(l/r) go through in the 'convert forall' pass.
The return value inside the	loop gets stored into a single
integer temp which does not need memory to be allocated,
thereby saving time and memory.