Support -Mallocatable=03 in device code
Generate RTE_allocated_lhs() when testing if the LHS of an allocatable
assignment is allocated when inside atarget compute region or in a routine
compiled for the device. This is equivalent to RTE_allocated() but allows the
accelerator code generator and runtime to provide special handling for this
case.

This is implemented by adding the ISASSIGNLHS flag to the I_ALLOCATED ast in
these circumstances. gen_allocated_check() and rewrite_deallocate() have a new
argument to indicate when the check is for the destination of an allocatable
assignment.

To correctly identify an assignment with STD_ACCEL set, the flag must be
preserved when it is commented out by ast_to_comment(). Otherwise, if
gen_allocated_check() occurs right before that STD, we wouldn't recognize it
as being in a compute region.

The code in lowerexp.c for lowering I_ALLOCATED was simplified but the only
functional change is to use the new routine when ISASSIGNLHS is set.

Rewrite the final gen_allocated_check() in rewrite_allocatable_assignment()
so that it is not associated with the next STD. It could be from the next
line, resulting in an incorrect line number or STD_ACCEL flag.