Support -Mallocatable=03 in device code
Generate RTE_allocated_lhs() when testing if the LHS of an allocatable
assignment is allocated when inside a target compute region or in a routine
compiled for the device. This is equivalent to RTE_allocated() but allows the
target code generator and runtime to provide special handling for this
case.

This is implemented by adding the ISASSIGNLHS flag to the I_ALLOCATED ast in
these circumstances. gen_allocated_check() and rewrite_deallocate() have a new
argument to indicate when the check is for the destination of an allocatable
assignment.

The code in lowerexp.c for lowering I_ALLOCATED was simplified but the only
functional change is to use the new routine when ISASSIGNLHS is set.