Fix an uninitialized memory reference to out of scope array descriptor
This occurs when running gfortran test alloc_dummy_1.

Reduced test case:

     1  program alloc_dummy
     2      integer, allocatable :: a(:)
     3      allocate(a(3))
     4      a = [1, 2, 3]
     5      if (all(whatever(a) == [1, 2, 3])) print*, "ok"
     6  contains
     7      function whatever(x)
     8          integer, allocatable :: x(:)
     9          integer :: whatever(size(x))
    10          whatever = x
    11      end function whatever
    12  end program alloc_dummy

For this test, code in host program alloc_dummy to set up storage
for the array value returned by the call to function whatever at
line 5 has an invalid access to descriptor x$sd, which is local to
whatever.  This occurs because the size of the array return value
depends on the size of dummy argument x.  flang1 routine
semfunc.c:replace_arguments must replace the reference to x$sd with
a reference to actual argument descriptor a$sd.  However, current
code incorrectly requires that a$sd must already exist at the time
replace_arguments is called.  Change replace_arguments code to
generate a$sd if it doesn't already exist.