Fix SELECT TYPE with unlimited polymorphic pointer to INTEGER(KIND=8)
In transform_call() of rest.c, in the case of a class(*) dummy
argument, we make sure the actual argument has its type information
set up prior to the call. If the actual argument does not normally
take a descriptor, we force it to have a descriptor argument since the
class(*) dummy argument requires a descriptor. In this case, we
also generate additional code for initializing the descriptor.
We do not want to generate the initialization code if the actual
argument normally has a descriptor since we can assume that its
descriptor was already been properly set up. In fact, unconditionally
generating this initialization code can corrupt the descriptor and
lead to memory errors in the user application. To prevent this
erroneous state, add a !needdescr guard, where needdescr is true
when the actual argument requires a descriptor.

Includes test case.