Correct handling of assumed-shape array arguments with TARGET attribute
Add xbit 58,0x400000 to not make a copy of assumed-shape array
arguments if the callee has the argument marked as target.

We do not make a copy and send a section, but send a pointer
to the array plus an array desciptor.

Initialize array descriptors with their type when the base
type is a derived type. We only do this for non-dummy arguments
and when they are first declared.