Add support for ILI representing tightly nested offloaded loops
Add a second field for the IL_ACCSLOOP, IL_ACCKLOOP, IL_ACCPLOOP
ILI, which is a short constant that will normally be zero,
but will be '1' when the loop clause was either on
serial/kernels/parallel constructs, or the loop construct is
tightly nested in serial/kernels/parallel constructs.

Add PR_ACCTKLOOP, PR_ACCTPLOOP, PR_ACCTSLOOP, which are used for
a loop directive or clause that is tightly nested (the T) inside
a kernels, parallel or serial constructs.