Fix ICE in default initialization of structure component
There are a couple issues here. First need to rewrite
"no_data_components" so that it handles recursive cycles both in
the front end and back end. In this case "no_data_components"
should return FALSE. Second, "all_default_init" should return
TRUE here, because the first and only component in the derived
type is default initialized to null. Modified "all_default_init"
so that it ignores compiler created symbols like first$p,
firest$o, first$sd. Modified dinit.c so that when the sptr
in dinit_val() is a pointer and val is 0 and dtypev is an
integer, do not call "cngcon", it should be just a straight
assignment. Since we don't set DINIT on individual members,
a new flag was created that marks the derived type as "all
default initialized". The front end passes in the flag to the
back end which will mark the derived type definition.

Fixes #240