Fix extent calculation wrapping at -O2
If the computation of "extent = upper - lower + 1" wraps, then checking
for extent < 0 won't work. Instead we need to check if upper < lower
to decide if extent to be set to 0.

This was fixed in the runtime in rdst.c, but _template in outconv.c does
the calculation in-line and had the same bug.