Optimize array bounds calculations on array expressions
When creating a forall loop for array assignment, if the LHS
array loop bounds are not constant, check RHS expression for
array with constant loop bound to see if we can use its shape
to make the forall loop. This will enable further optimization.
This submission allows simple one dimension array expression.
More work can be done on complex expresion or multiple rank
arrays in the future.