Improve handling of bounds variables for adjustable array return values
When a contained subroutine returns an adjustable array and its
bounds is the host routine variable, its ast symbol will not be
exported back to host routine.  Reuse IDSTR flag which has been
used for character type so that we can export symbol instead
of ast and during import if it sees that the symbol IDSPTR flag
is set, we will check this flag and get symbol instead of ast.