Fastmath cleanup
Currently, there are multiple paths through directives.h depending
on which CPP macros are defined.  Specifically, there are certain
objects that are built - for example using the LINUX_ELF section of
definitions, while there are other objects that are built using the
"default" section of macro definitions.

This change will:
1) Standardize the use of CPP macros TARGET_LINUX_X8664,
   TARGET_OSX_X8664, and TARGET_WIN_X8664 for all libpgm objects.

   Also use same set of CPP macros to build the assembly based objects
   in libcpuid.

2) Remove all unused architectures/platforms from directives.h
   For example: INTERIX86, OSF.

3) Remove CPP macros MOVDQ and VMOVDQ which were mapped to movd and
   vmovd respectively.

4) Continue to simplify if/elif/.../endif logic in determining
   X8664 platforms in directives.h.