Add support for transformational intrinsics in constant contexts
Enables tranformational intrinsic support for MAXVAL, MAXLOC,
MINVAL, and MINLOC in contexts like parameter definition and
array bounds.

module foo
  INTEGER, PARAMETER :: nh2oref(2) = (/10, 4/)! #self, #foreign
  INTEGER, DIMENSION(MAXVAL(nh2oref)) :: h2oref_in
end module foo

This is implemented in both the front-end and back-end.

Previously, compiler fails in compilation when using intrinsics
for an array bounds definition inside a module. The problem is
that compiler uses an adjustable array type for array h2oref_in
which fails in module syntax checking. So we need to enable
expression evaluation for intrinsic arguments, and use constant
dtype when we find a constant value generated. This fix sets
conditions to return TRUE for ast_isparam when processing
MAXVAL, MAXLOC, MINVAL, and MINLOC. It also fixes the issue
with handling subroutine parameter arrays passed as arguments
to MAXVAL and MINVAL.