Allow the vectorizer to rewrite more loops from OMP SIMD
Snapshot of passes to rewrite the ILI to allow the LLVM vectorizer
to work on more loops. The passes are under the control of the xbit
183 0x10000000.

The goal is to allow loops marked with "#pragma omp simd" to be
vectorized by the LLVM Vectorizer for performance.

Rewrite of the widening pass. Makes all address arithmetic operations
the same width and eliminating sign-extensions and truncations.

Adding load-load eliminating pass. Cache unneeded indirections from
passing by reference.