Resolve ambiguity with resolving overloaded function
In tkr_match() of semgnr.c, handle the case where the actual argument
is an identifier to a procedure, but the dummy argument (called
formal in tkr_match()) is not. In this case, we need to return
INF_DISTANCE. Otherwise, the actual argument's return type will
incorrectly match with the dummy argument's type. This can lead to
ambiguous matches in the generic procedure/interface resolution.