Fix improper handling of the '\r' character in strings.
This fixes a bug with our handling of the '\r'
character in a string. Characters such as '\n' and '\r' need
special handling. We handle '\n' correctly, but not '\r'.
The '\r' character is now handled in the same way as '\n' in
the ll_get_string_buf() functon of source file llutil.c.