Fix a bug with iso_c_binding and privatization.
This fixes Flang issue #366. When adding the "private" attribute to
iso_c_binding instrinsic symbols, we were getting an error such as the
following:

"illegal use of symbol c_loc/c_f_pointer/... - attempt to CALL a
non-SUBROUTINE".

The Problem is that when we process the private statement, we create a forward
reference symbol with a symbol table type (stype) of  ST_UNKNOWN. The
do_access() function would then change the ST_UNKNOWN stype to ST_MODPROC which
is wrong in this case. We need to be referencing the ST_INTRIN symbol instead.

To solve this issue, we call findByNameStypeScope() to find the symbol
with with an stype of ST_INTRIN and let the ST_UNKNOWN symbol act as an alias
of the ST_INTRIN symbol (i.e., the ST_UNKNOWN stype becomes ST_ALIAS and its
SYMLK field points to the ST_INTRIN symbol).