Fix a bug with procedure variable/pointer interfaces.
In the  _do_iface() and defer_iface() functions,
keep track of gbl.internal of the procedure variable/pointer that
we are processing.

In _do_iface(), if the procedure variable/pointer was declared
in an internal procedure that differs from the current procedure,
then we need to skip it. Otherwise, we may incorrectly overwrite the
procedure variable/pointer's dtype.

In semant.h, add internal field to IFACE struct. Also add doxygen
to the IFACE typedef.