Fix an ICE that was caused by incorrectly processing type descriptor arguments.
The Problem is that we are passing 0 to mk_id() in transform_call() of rest.c.
The sptr is a type descriptor argument, so its descriptor field is 0.
The fix is to add a guard to prevent making a descriptor argument from a
descriptor (which lead to the 0 argument to mk_id()).