Fix Flang issue #467.
Fix the error of "use of undefined value ..." for Flang when we have a function
that returns an object of c_ptr type.

Flang previously used C calling convention with functions that return an object of c_ptr type. However, this was changed to the Flang calling convention
(e.g., c_ptr results are now returned in the first argument register) when
BIND(C) is not used on the procedure. This is needed to properly support
Fortran specific features such as type bound procedures that have a c_ptr
result.

Note: The Fortran specification requires BIND(C) if C calling convention is
required. Therefore, switching to the Flang calling convention is legal
(and correct) when BIND(C) is not used.