Fix Flang issue #446.
In cmp_interfaces_strict() of symtab.c, add the ability to relax the
checking of the PURE attribute between two interface arguments (e.g.,
dummy and actual arguments). For example, it is legal to
pass a "pure procedure" to a procedure dummy argument that does not
have the pure attribute (provided all of the rules for passing a
procedure to a procedure dummy argument apply). To support this
ability add 2 flags for cmp_interface_strict() that tell the function to
relax the "pure check".

Note: The ability to strictly check the pure and impure flags is still
required when checking type bound procedure interfaces between
parent type and type extensions (i.e., betwen parent and child type
declarations).

In symtab.in.h, add RELAX_PURE_CHK_1 and RELAX_PURE_CHK_2 flags to
relax the pure check.

RELAX_PURE_CHK_1 tells cmp_interfaces_strict() not to apply the pure
check on its first interface argument (sym1). That is, it's OK that
sym1 has the PURE attribute. RELAX_PURE_CHK_2 is the same as
RELAX_PURE_CHK_1, but it applies to the sym2 argument.

We relax the pure check in calls to cmp_interfaces_strict() in
transform_call() of rest.c and in compat_arg_lists() of semfunc2.c.