Fix a bug related to move_alloc().
In transform_move_alloc() of func.c, remove the call to
move_alloc_type().
Also remove the move_alloc_type() function since it is no longer used.

When applicable, this function was used to assign the dynamic
type of move_alloc()'s "from" argument to its "to" argument. However,
it turns out that the move_alloc() runtime routine also performed this
same assignment, so the assignment in move_alloc_type() was redundant
and sometimes wrong.

For example, it did not correctly access the dynamic type from a
a descriptor that's embedded in a derived type. The move_alloc()
runtime routine properly supports this case.