Cleanup LIBPGMATH for all architectures/platforms.
In the creation of the the original libpgm/libpgmath, there are some
objects that were included in the library that in hindsight do not
belong there.  Without loss of generality they fall into two
categories:

1) Entry points prefixed with "__mth_i" that do not fit the defintion
   of "math" vis a vis the intent of libpgmath.

2) "Math" routines that are prefixed with "ftn_".  Again those do not
   meet the intent of libpgmath.

Objects are better suited to be included in the Flang runtime
library(ies).

This change also includes moving the KMIN/KMAX objects into
the Flang runtime libraries.