Fix a bug with pointer/allocatable variables in modules with save statements.
We create descriptors for pointers and allocatables. If they are
a basic type (i.e., not arrays, derived types, nor polymorphic), then
they are simple variables. In most of these cases, we can use a one word
"pseudo descriptor". A change was made in transform_call()
that assumed regular descriptors in this case. There is usually not
a problem with this, except in the case where a module has a save
statement. This places these descriptor variables in BSS (static)
memory. Unfortunately, the BSS section may not be visible by the
user of the module. Therefore, to eliminate this problem, we will
go back to using the "pseudo descriptor" in transform_call() for
basic typed objects (since they can be created on the fly).

Note: This problem does not apply to non-basic typed objects such as
arrays, derived types, and polymorphic objects because their
descriptor is either embedded in a derived type (if they are derived
type components) or they are created in the user scope.