Add an enhancement to the array_element_dtype() function of dtypeutl.c.
--- dtypeutl.c
Change the array_element_dtype(dtype) function to return dtype if the dtype is
not an array dtype.

The original behavior was to return DT_NONE, but all the existing calls
to array_element_dtype() first tested if it was an array dtype.

--- dtypeutl.h
Fill in the \brief doxygen for the array_element_dtype() function.