Fix a bug related to printing empty structure constructors.
We did not handle the case where we printed derived types with empty structure
constructors. Assigning the results to a temporary variable and printing it
will work. For this case, we should print the derived type components one by
one. The fix is to create a temporary variable, initialize it with value of
the structure constructor, and then print the temporary variable.