Cleanup some code in the Flang back-end.
- get rid of redundant debug macros
- start fixing dinit.c for type correctness

Split the dinitutl.h and dinit.h files and clean up the modules
that depend on them.

- Prune the number of include files
- Add forward declarations for pointer-to-struct types
- Get rid of redundant prototypes

Changes that allow direct.c to be compiled as C++ sans errors
or warnings.

Set of changes to start introducing type correct accessor functions
to the Fortran backend.

Cleanup the symtab.in.h file.
- get rid of misplaced prototypes
- forces .c files to include the correct .h files
- cleanup other missing include files.