Fix a bug that caused unresolved type descriptor symbols.
Do not create type descriptors for derived types inside interfaces.
When derived types are defined inside interfaces, type descriptors
are not needed because there is no executable code inside the
interface. Futhermore, if we generate them, we might get multiple
definitions of the same type descriptor because we mangle the name
based on the scope of the interface; not on the scope of the host
routine.

To implement this solution, we add a symbol table flag called
IS_INTERFACE to ST_ENTRY and ST_PROC that is set if the symbol
is in an interface block. This is set in semant1. We then check
this flag in lower_symbols() to determine if we need to lower
a particular type descriptor. If the type_descriptor's scope
is inside an interface, we do not lower it. By not lowering the
type descriptor, its generation in the back-end does not occur.