Address two issues involving poly_asn calls.
The first issue is a regression in a test.  The problem is an
incorrect poly_asn call for a conformable assignment of a scalar
to an array.  The fix for this problem is to check for this case in
routine semutil2.c:_construct90.  The remaining code changes address
the second issue.

The second issue is an incorrect result for code that does a sourced
allocation of an unlimited polymorphic object.

    class(*), allocatable :: val(:)
    call get_value (val)
    select type (val)
    type is (character(*))
      print '(2a)', val
    end select
    contains
      subroutine get_value (value)
        class(*), allocatable, intent(out) :: value(:)
        allocate(value, source=['one','two'])
      end subroutine
    end

Output for this code should be 'onetwo', not 'on'.  The fix for this
problem requires both front end and runtime code changes.  Runtime
code defines two additional routines related to poly_asn:

  poly_asn_src_intrin
  poly_asn_dest_intrin

These routines are wrapper routines for special case calls to
poly_asn where either the source or destination of the
assignment are intrinsic types rather than derived types.
Although these routines are defined, they are not currently used.
The fix is to arrange for these routines to be available for use
in the front end, and call poly_asn_src_intrin instead of
poly_asn when the source is an (intrinsic) character object.
The runtime code for poly_asn must also be updated to make
use of this information.  Note that the character case is unique
in that it is the only intrinsic type that needs a length, which
appears in a different runtime location than other date type sizes.
It may be that other intrinsic types can manage with (somewhat
awkward) direct calls to poly_asn, so it isn't immediately
clear if other intrinsic type cases should use either
poly_asn_src_intrin or poly_asn_dest_intrin.  However,
it should now be possible to do that if it becomes necessary.