Fix a bug related to a conformable_dd() runtime call.
The conformable_dd() runtime call has an invalid argument

Under F2003 allocation semantics (now the default), an assignment 'c = v' where

 - c is an allocatable array; and
 - v is a dummy argument array with one or more non-constant bounds

generates a call to conformable_dd() with an invalid descriptor
for v as part of the deallocation/reallocation sequence for array c.
(The two d's in conformable_dd() denote destination and source
descriptors.)  This conformable_dd() call is generated in routine
transfrm.c:build_conformable_func_node.  It isn't clear why code
prior to build_conformable_func_node links primary object v with an
associated descriptor without actually building a valid descriptor.
It may be that other code paths do arrange to create a valid descriptor
at some point before it is used.  However, for the case at hand,
information about array v is available without reference to a
descriptor.  For this test case, it should be at least as efficient
at runtime to use that information directly rather than packaging it in
a descriptor.  So, the fix is to use this information to replace the
"dd" call with routine conformable_d1v(), which passes the array's
extent, rather than a descriptor.  With this change, other cases
such as those referencing multi-dimension arrays might choose a
different non-descriptor runtime routine to call.