Fix a bug with an initial class declaration in an anonymous program unit.
Flang compilations of an anonymous program where the
first statement is a class declaration fail with an undefined reference
to a compiler generated type descriptor variable because the compiler
expects type descriptor variables to begin with a program unit name
prefix.  This expectation fails for the compiler because the front
end doesn't account for the possibility that the class declaration is
in an anonymous program unit until after the initial class declaration
is processed.

A provisional attempt to solve the problem by relaxing the flang back
end to remove the prefix requirement appeared to be successful
(llassem.c:write_typedescs).  However, the fortran grammar defines a
null production to provide a mechanism for accounting for cases like
this, which is used in 30+ productions.  So, the preferred fix is to
modify the grammar to apply this mechanism to class declarations.
The result is a type descriptor object that has a "MAIN" prefix that
all back ends can process.

Also add oop731 f90_correct test to exercise this bug fix.