Fix 64-bit CHARACTER Length in runtime.
This change replaces all uses of CLEN/DCLEN from "int" type to "size_t"
type. We created a new __CLEN_T macro for this purpose. There
should no longer be any bare "int" variables that store a CHARACTER
length, or have a CHARACTER length passed into a function received as
an "int". This will now be much easier to update in the future, should
we have need.

Most of the changes are in the FORTRAN runtime. Intrinsic routines that
used to take a 32-bit CHARACTER length have been updated to the
__CLEN_T type. In order to keep backwards compatibility we need to keep
the old interface, and create a new routine with a new interface. The
convention chosen is to append an "a" to the new routine. E.g.,
ftn_len(... DCLEN()) becomes ftn_lena(... CLEN64()). Future updates
will allow for ftn_lenb(), etc. The new routine contains the
definition, and the old interface calls the new interface by casting
its 32 bit type to 64 bit. All of the old interfaces have been marked
with "/* 32 bit CLEN version */", since they are deprecated. We can
remove these routines in the future.

FuncArgMacros.h
Add the __CLEN_T and DCLEN64 macros.

rtlRtns.[ch]
All functions that have been updated and are generated by the compiler
have had their entry updated to reflect their new name.