Fix a bug related to intrinsic and variable name overloading.
A test case with an intrinsic routine reference in a host routine
that also has contained routine that has a variable with the same
name as the intrinsic fails to compile.  The fix is to check for
an explicit INTRINSIC attribute when processing the contained
routine variable to determine if the two references refer to the
same symbol.

Also add a test to exercise the bug fix.