libpgmath improvements Add pure AVX513 single and double precision implementations of elemental functions ACOS(), ASIN(), EXP(), LOG(), POW(), and ATAN(). X86-64 Only.
One of the primary objectives of this commit is to, as much as
possible, maintain source code compatibility with the scalar
(32/64-bit), 128-bit, and 256-bit versions.  The primary implication
is that the AVX512 mask register is not explicity referenced in the
source code.  Compatibility with XMM and YMM verions is maintained
by implementing AVX512 versions of the X/YMM compare and merge
operations.  This implementation choice provids the fastest and
cleanest way to get from 256-bit to 512-bit while only incurring a
minor performance loss.

This commit will enable the new FMA3 AVX512 versions of these functions
in the dispatch configuration tables.

Also fix FMA3 POW(base,1.0) not being equal to base.