Fix structure constructor with CLASS(*) allocatable component
With an allocatable unlimited polymorphic member in a
derived type, using the structure constructor does not assign the
right dynamic data type to the member.  There are two problems
here. One is that space isn't allocated. The other is that the
right type isn't being assigned.

This addresses part of Flang issue #243.