Correct handling of overloaded associate names
Symptom is error messge "insert_sym: bad hash"

In construct_association(), call insert_sym_first() instead of
insert_sym() to make sure an overloaded associate name is
accessible in the symbol table. For example, if the associate
name is the same as the function we are currently processing,
we want to make sure we get the associate symbol; not the
function symbol which is a forward reference and incomplete
when we process the select type.