Correct argument passing with character lengths and procedure pointers
When we have character length arguments, we shift any procedure
dummy argument descriptors to the end of the argument list.
We accomplish this in part through a symbol table flag called
IS_PROC_DESCR which is set on any procedure dummy argument
descriptors.  After introducing the aforementioned policy in
a previous source update, a couple new bugs were introduced.

Bug #1: When compiling with -Minline, we lost the IS_PROC_DESCR
flag on any inlined procedure dummy argument descriptors.
Therefore, we need to preserve this flag during the inliner's
extract phase (see extract.c and inliner.c changes).

Bug #2: When we mix character length arguments, procedure dummy
arguments, and optional arguments, we cannot safely shift the
procedure dummy arguments. Fortunately, this is not much of a
limitation because in most cases the user is going to use an
interface for the procedure that gets called.  They could still
declare the procedure as external for the called procedure,
but external by definition cannot have holes in its argument
list. So, this limitation does not really affect external
either. We still want to shift the arguments in all other
cases since the called procedure could be a ``C'' (or ``C''
like) routine.