Fix issues with new uplevel; format consistency changes
With new uplevel method, an uplevel structure is created on
the stack and pass to nested task routine.  This may not work
as the task may get executed after its task caller exit and
anything put on stack will be invalid.  Hence, copy address
of all shared variables on task data allocated by runtime.

The compiler also creates a pointer of DT_ANY. cgmain should
allow that dtype if it uses dt_nme to check for alignment.