Fix segmentation fault due to uninitialized field in subc pointer
In get_minmax_val(), after allocating CONST*temp using
getitem(), initialize the newly allocated memory to zero, so
that all fields are initialized, especially all pointer fields
are NULL. Before this, some tests were failing with a segfault
because it was following an uninitialized 'subc' pointer.

Also: fix a spelling error in an interr() message.