Ensure contiguity of data passed to __kmpc_threadprivate_cached
When we pass the address of the master thread for copyin,
its address must always be from the result of
__kmpc_threadprivate_cached from an outer region.

For an implicit save threadprivate (local in main program), make
the variable save for pointer/allocatable.
__kmpc_threadprivate_cached always copies from the address we
pass in up to the size we pass. The size include size of $p $sd $o.
We always pass address of $p (we assume that it is the beginning
of original data).

By making the variable "SAVE", they will be next to each other
or at least $p is at the beginning of them all and we won't point
to an invalid address when pass to kmpc runtime.