Correct erroneous elimination of INT4 to INT8 type conversion
The compiler had a KSUB ILM take an ICON ILM operand
instead of a KCON ILM operand.

This occurred in the A_CONV case of lower_ast() in lowerexp.c.
The compiler incorrectly eliminated the type conversion (from INT4 to INT8)
when there was a constant expression without first checking whether the
type of the constant and type of the conversion (A_CONV) was the same.