Fix a bug generating an incorrect "virtual function table"
This bug occurred when there was an abstract parent type in the
same module as a child type that defines one or more deferred type
bound procedures. When these type bound procedures are defined,
the wrong offset was getting into the child type's "virtual function
table". To fix this problem, make sure to use the parent's offset
for these functions when defining the child's functions.