Generate correct LLVM IR when using "unreachable" instructions
A recent change to implement __builtin_unreachable() with an LLVM
"unreachable" instruction caused a regression.  The "unreachable"
instruction, which is a terminal instruction, was not followed by a
label, which is illegal LLVM IR and caused llc to fail with an error.

Fix the regression by forcing a label after an "unreachable"
instruction, and by being smarter about which instructions are
terminal instructions.