Fix compiler internal error in lowering with F2003 allocatable semantics
In rewrite_allocatable_assignment() when generating a call
to poly_asn(), simplify the source argument when it is
an A_SUBSCR ast. To accomplish this, call a new function
called mk_ptr_subscr() which returns the array A_ID ast
when it's an empty subscript (i.e., references the entire
array). Otherwise, return a pointer to a contiguous shallow
copy of the array slice.  The poly_asn() function will then
perform any necessary deep copies of the array elements. Note
also that the destination array inherits its shape from the
source array, so this A_SUBSCR issue does not occur with
the destination array. However, if the destination AST is an
A_SUBSCR, just pass in the array's A_ID ast.

Also add masks for the empty parts of the subscript expression
(i.e., lboundMask, uboundMask, and strideMask).

Replace hard coded 0x7 mask value with the aforementioned
named constants.